+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0](https://github.com/csipiemonte/vide-pwa/compare/v1.0.0...v1.1.0) - 2020-12-11
+
+### Added
+- Controllo attivazione audio/video.
+- Test e2e.
+
+### Changed
+- Fix scaricamento file in dettaglio appuntamento.
+
 ## [1.0.0](https://github.com/csipiemonte/vide-pwa/releases/tag/v1.0.0) - 2020-10-27
 
 ### Versione iniziale